refactor(hog-charts): move BarChart into its own folder with utils
Mirrors the `TimeSeriesLineChart/` layout. New structure:

  charts/BarChart/
    BarChart.tsx
    BarChart.test.tsx
    BarChart.stories.tsx
    BarTooltip.tsx           — tooltip wrapper that narrows seriesData
                               to the cursor's bar in grouped layout
    utils/
      bars-under-cursor.ts   — `barContainsPoint` + `seriesKeysAtCursor`,
                               shared by drawHover and BarTooltip
