+++ frontend/src/lib/hog-charts/charts/BarChart/BarChart.tsx
@@ -1,18 +1,16 @@
 import React, { useCallback, useMemo } from 'react'
 
-import { type BarChartPrivate, computeBarAtIndex, computeSeriesBars } from '../core/bar-layout'
-import { type BarRect, drawBarHighlight, drawBars, drawGrid, type DrawContext } from '../core/canvas-renderer'
-import { Chart } from '../core/Chart'
-import { useChartLayout } from '../core/chart-context'
-import { ChartErrorBoundary } from '../core/ChartErrorBoundary'
+import { type BarChartPrivate, computeBarAtIndex, computeSeriesBars } from '../../core/bar-layout'
+import { type BarRect, drawBarHighlight, drawBars, drawGrid, type DrawContext } from '../../core/canvas-renderer'
+import { Chart } from '../../core/Chart'
+import { ChartErrorBoundary } from '../../core/ChartErrorBoundary'
 import {
-    type BarScaleSet,
     computePercentStackData,
     computeStackData,
     createBarScales,
     type StackedBand,
     yTickCountForHeight,
-} from '../core/scales'
+} from '../../core/scales'
 import type {
     BarChartConfig,
     ChartDimensions,
@@ -24,58 +22,15 @@ import type {
     ResolvedSeries,
     Series,
     TooltipContext,
-} from '../core/types'
-import { DEFAULT_Y_AXIS_ID } from '../core/types'
-import { computeVisibleXLabels } from '../overlays/AxisLabels'
-import { DefaultTooltip } from '../overlays/DefaultTooltip'
+} from '../../core/types'
+import { DEFAULT_Y_AXIS_ID } from '../../core/types'
+import { computeVisibleXLabels } from '../../overlays/AxisLabels'
+import { BarTooltip } from './BarTooltip'
+import { seriesKeysAtCursor } from './utils/bars-under-cursor'
 
 function bandCenter(scales: BarChartPrivate['__barChart'], label: string): number | undefined {
     const start = scales.band(label)
     return start == null ? undefined : start + scales.band.bandwidth() / 2
-}
-
-function barContainsPoint(bar: BarRect, point: { x: number; y: number }): boolean {
-    return point.x >= bar.x && point.x <= bar.x + bar.width && point.y >= bar.y && point.y <= bar.y + bar.height
-}
-
-interface BarHitTestArgs {
-    series: readonly Pick<Series, 'key' | 'visibility' | 'yAxisId' | 'data'>[]
-    label: string
-    dataIndex: number
-    cursor: { x: number; y: number }
-    scales: BarScaleSet
-    layout: 'stacked' | 'grouped' | 'percent'
-    isHorizontal: boolean
-    stackedData?: Map<string, StackedBand>
-    topStackedKeyByAxis: Map<string, string>
-}
-
-/** Shared by drawHover and the tooltip wrapper so they can't drift. */
-function seriesKeysAtCursor(args: BarHitTestArgs): Set<string> {
-    const { series, label, dataIndex, cursor, scales, layout, isHorizontal, stackedData, topStackedKeyByAxis } = args
-    const hits = new Set<string>()
-    for (const s of series) {
-        if (s.visibility?.excluded) {
-            continue
-        }
-        const stackedBand = stackedData?.get(s.key)
-        const axisId = s.yAxisId ?? DEFAULT_Y_AXIS_ID
-        const isTopOfStack = topStackedKeyByAxis.get(axisId) === s.key
-        const bar = computeBarAtIndex({
-            series: s as Series,
-            label,
-            dataIndex,
-            scales,
-            layout,
-            isHorizontal,
-            stackedBand,
-            isTopOfStack,
-        })
-        if (bar && barContainsPoint(bar, cursor)) {
-            hits.add(s.key)
-        }
-    }
-    return hits
 }
 
 export interface BarChartProps<Meta = unknown> {
@@ -331,20 +286,6 @@ function BarChartInner<Meta = unknown>({
         [stackedData, barLayout, isHorizontal, topStackedKeyByAxis, barCornerRadius]
     )
 
-    const wrappedTooltip = useCallback(
-        (ctx: TooltipContext<Meta>): React.ReactNode => (
-            <BarTooltipShell<Meta>
-                ctx={ctx}
-                userTooltip={tooltip}
-                stackedData={stackedData}
-                topStackedKeyByAxis={topStackedKeyByAxis}
-                layout={barLayout}
-                isHorizontal={isHorizontal}
-            />
-        ),
-        [tooltip, stackedData, topStackedKeyByAxis, barLayout, isHorizontal]
-    )
-
     const resolveValue = useMemo(() => {
         if (!stackedData) {
             return undefined
@@ -369,7 +310,16 @@ function BarChartInner<Meta = unknown>({
             createScales={createScales}
             drawStatic={drawStatic}
             drawHover={drawHover}
-            tooltip={wrappedTooltip}
+            tooltip={(ctx) => (
+                <BarTooltip<Meta>
+                    ctx={ctx}
+                    userTooltip={tooltip}
+                    stackedData={stackedData}
+                    topStackedKeyByAxis={topStackedKeyByAxis}
+                    layout={barLayout}
+                    isHorizontal={isHorizontal}
+                />
+            )}
             onPointClick={onPointClick}
             className={className}
             dataAttr={dataAttr}
@@ -378,61 +328,4 @@ function BarChartInner<Meta = unknown>({
             {children}
         </Chart>
     )
-}
-
-interface BarTooltipShellProps<Meta> {
-    ctx: TooltipContext<Meta>
-    userTooltip?: (ctx: TooltipContext<Meta>) => React.ReactNode
-    stackedData: Map<string, StackedBand> | undefined
-    topStackedKeyByAxis: Map<string, string>
-    layout: 'stacked' | 'grouped' | 'percent'
-    isHorizontal: boolean
-}
-
-function BarTooltipShell<Meta>({
-    ctx,
-    userTooltip,
-    stackedData,
-    topStackedKeyByAxis,
-    layout,
-    isHorizontal,
-}: BarTooltipShellProps<Meta>): React.ReactElement {
-    const { scales } = useChartLayout()
-    const d3Scales = (scales._private as BarChartPrivate | undefined)?.__barChart
-    // Stacked/percent: keep all rows so the tooltip mirrors the whole hovered column —
-    // matches the legacy chart.js bar tooltip's default `mode: 'nearest'` for stacked bars.
-    const narrowed =
-        layout === 'grouped' && d3Scales && ctx.hoverPosition && ctx.dataIndex >= 0
-            ? narrowSeriesByCursor(ctx, d3Scales, layout, isHorizontal, stackedData, topStackedKeyByAxis)
-            : ctx
-    return <>{userTooltip ? userTooltip(narrowed) : DefaultTooltip(narrowed)}</>
-}
-
-function narrowSeriesByCursor<Meta>(
-    ctx: TooltipContext<Meta>,
-    scales: BarScaleSet,
-    layout: 'stacked' | 'grouped' | 'percent',
-    isHorizontal: boolean,
-    stackedData: Map<string, StackedBand> | undefined,
-    topStackedKeyByAxis: Map<string, string>
-): TooltipContext<Meta> {
-    const cursor = ctx.hoverPosition
-    if (!cursor) {
-        return ctx
-    }
-    const hits = seriesKeysAtCursor({
-        series: ctx.seriesData.map((entry) => entry.series),
-        label: ctx.label,
-        dataIndex: ctx.dataIndex,
-        cursor,
-        scales,
-        layout,
-        isHorizontal,
-        stackedData,
-        topStackedKeyByAxis,
-    })
-    if (hits.size === 0) {
-        return ctx
-    }
-    return { ...ctx, seriesData: ctx.seriesData.filter((entry) => hits.has(entry.series.key)) }
 }
