+++ frontend/src/lib/hog-charts/index.ts
@@ -1,6 +1,6 @@
 // Components
-export { BarChart } from './charts/BarChart'
-export type { BarChartProps } from './charts/BarChart'
+export { BarChart } from './charts/BarChart/BarChart'
+export type { BarChartProps } from './charts/BarChart/BarChart'
 export { LineChart } from './charts/LineChart'
 export type { LineChartProps } from './charts/LineChart'
 export { TimeSeriesLineChart } from './charts/TimeSeriesLineChart/TimeSeriesLineChart'
